Complete refactor to move thermometer dynamic svg to a component, and abstract temp colours to their own const file.
