+++ web-ui/src/constants/temperatureColours.ts
@@ -1,0 +1,7 @@
+export const temperatureColours = {
+  subZero: "#00B7C3",
+  cold: "#00CC6A",
+  warm: "#FFB900",
+  hot: "#F7630C",
+  veryHot: "#E81123",
+};
